# almacen2/api_doc.yaml
openapi: "3.0.2"
info:
  title: Almacén API
  description: API para manejo de almacén
  version: 1.0.0
servers:
  - url: http://192.168.77.29:5000
    description: Servidor local
paths:
  /articles:
    post:
      summary: Crea un artículo en el almacén
      tags: 
        - Articles
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Información del artículo a crear
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                article_info:
                  type: string
                units:
                  type: integer
                available:
                  type: boolean
              required:
                - article_info
                - units
                - available
      responses:
        '201':
          description: Article created successfully
        '401':
          description: API Key is missing or invalid
    get:
      summary: Obtiene la lista de todos los artículos del almacén
      tags:
        - Articles
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    article_info:
                      type: string
                    units:
                      type: integer
                    available:
                      type: boolean
        '401':
          description: API Key is missing or invalid
    
  /articles/{article_id}:
    get:
      summary: Obtiene un artículo concreto del almacén
      tags:
        - Articles
      parameters:
        - in: path
          name: article_id
          description: ID del artículo a obtener
          required: true
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  article_info:
                    type: string
                  units:
                    type: integer
                  available:
                    type: boolean
        '401':
          description: API Key is missing or invalid

    put:
      summary: Actualiza un artículo existente en el almacén
      tags:
        - Articles
      parameters:
        - in: path
          name: article_id
          description: ID del artículo a actualizar
          required: true
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        description: Información del artículo a actualizar
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                article_info:
                  type: string
                units:
                  type: integer
                available:
                  type: boolean
              required:
                - article_info
                - units
                - available
      responses:
        '200':
          description: Article updated successfully
        '401':
          description: API Key is missing or invalid
    delete:
      summary: Elimina un artículo existente en el almacén
      tags:
        - Articles
      parameters:
        - in: path
          name: article_id
          description: ID del artículo a eliminar
          required: true
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Article deleted successfully
        '401':
          description: API Key is missing or invalid

  /articles/receive/{article_id}:
    put:
      summary: Recibir un artículo de almacén
      tags: 
        - Articles
      parameters:
        - name: article_id
          in: path
          description: ID del artículo que se va a recibir
          required: true
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                units_received:
                  type: integer
              required:
                - units_received
      responses:
        '200':
          description: Article received successfully
        '401':
          description: API Key is missing or invalid
  
  /articles/send/{article_id}:
    put:
      summary: Enviar un artículo del almacén
      tags: 
        - Articles
      parameters:
        - name: article_id
          in: path
          description: ID del artículo que se va a enviar
          required: true
          schema:
            type: integer
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cantidad:
                  type: integer
              required:
                - cantidad
      responses:
        '200':
          description: API Key is missing or invalid
        '401':
          description: Unauthorized
  /consumers:
    post:
      summary: Crear un consumidor nuevo
      tags: 
        - Consumers
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Consumer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  api_key:
                    type: string
        '401':
          description: API Key is missing or invalid
  /consumers/authenticate:
    post:
      summary: Autenticar consumidor
      tags: 
        - Consumers
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  api_key:
                    type: string
        '401':
          description: API Key is missing or invalid